requirements.txt
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
bleach==4.1.0
cffi==1.15.0
colorama==0.4.4
debugpy==1.6.0
decorator==5.1.1
defusedxml==1.7.1
entrypoints==0.4
executing==0.8.3
ipykernel==6.9.2
ipython==8.2.0
ipython-genutils==0.2.0
jedi==0.18.1
Jinja2==3.1.1
jsonschema==4.4.0
jupyter-client==7.1.2
jupyter-core==4.9.2
jupyterlab-pygments==0.1.2
MarkupSafe==2.1.1
matplotlib-inline==0.1.3
mistune==0.8.4
mysql-connector-python==8.0.28
nbclient==0.5.13
nbconvert==6.4.5
nbformat==5.2.0
nest-asyncio==1.5.4
notebook==6.4.10
numpy==1.22.3
packaging==21.3
pandas==1.4.1
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
prometheus-client==0.13.1
prompt-toolkit==3.0.28
psutil==5.9.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
Pygments==2.11.2
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
pytz==2022.1
pywin32==303
pywinpty==2.0.5
pyzmq==22.3.0
scipy==1.8.0
Send2Trash==1.8.0
six==1.16.0
soupsieve==2.3.1
stack-data==0.2.0
terminado==0.13.3
testpath==0.6.0
toml==0.10.2
tornado==6.1
traitlets==5.1.1
wcwidth==0.2.5
webencodings==0.5.1
